✨ Tightly packed FeeParams Struct

+++ src/core/MarketCurve.sol
@@ -156,7 +156,7 @@ contract MarketCurve {
         }
 
         // Retrieve fee parameters from the fee manager contract
-        (address feeTo, uint256 BASIS_POINTS,, uint256 tradeFee,) = mom.feeParams();
+        (address feeTo, uint16 BASIS_POINTS, uint16 tradeFee,,) = mom.feeParams();
 
         // Calculate the trade fee based on the input amount
         uint256 fee = (xIn * tradeFee) / BASIS_POINTS;
@@ -211,7 +211,7 @@ contract MarketCurve {
         out = quote;
 
         // Get fee parameters from the mom contract
-        (address feeTo, uint256 BASIS_POINTS,, uint256 tradeFee,) = mom.feeParams();
+        (address feeTo, uint16 BASIS_POINTS, uint16 tradeFee,,) = mom.feeParams();
         // Calculate the trade fee
         uint256 fee = (out * tradeFee) / BASIS_POINTS;
         // Deduct the trade fee from the output amount
@@ -253,7 +253,7 @@ contract MarketCurve {
         status = Status.Graduated;
 
         // Get the fee parameters from the MarketFactory
-        (address feeTo,,,, uint256 graduationFee) = mom.feeParams();
+        (address feeTo,,,, uint128 graduationFee) = mom.feeParams();
         uint256 xToLP = balances.x - graduationFee;
 
         // Send the graduation fee to the specified fee recipient
